requirements.txt
flask==3.0.0
PyPDF2==3.0.1
fastembed==0.3.5
faiss-cpu>=1.7.4
numpy>=1.26.0
python-dotenv==1.0.0
flask-cors==4.0.0
reportlab>=4.0.0
setuptools>=68.0.0
wheel>=0.40.0
